Refactored major portions of the code
